+++ code/cut_separators/cand_callback.cpp
@@ -151,12 +151,24 @@ void callbacks::candidate_callback(CPXCALLBACKCONTEXTptr context, const hplus::e
     // exec.cand_cuts is a string containing one letter per type of cut to be applied
     // -> f : frontier landmark cuts
     // -> c : complementary landmark cuts
+    // -> l : LMcut cuts
     // -> s : SEC
-    if (exec.cand_cuts.find('f') != std::string::npos) usercuts_lm += cand_cuts::add_front_lm_cut(context, inst, unused_actions, reachable_state);
-    if (exec.cand_cuts.find('c') != std::string::npos)
+    if (exec.cand_cuts.find('f') != std::string::npos) {
+        usercuts_lm += cand_cuts::add_front_lm_cut(context, inst, unused_actions, reachable_state);
+    }
+    unsigned int lmc_violated{0};
+    if (exec.cand_cuts.find('l') != std::string::npos) {
+        lmc_violated = cand_cuts::add_lmcut_lm_cut(context, inst, unused_actions);
+    }
+    usercuts_lm += lmc_violated;
+    if (exec.cand_cuts.find('c') != std::string::npos || lmc_violated == 0) {
+        // Since the lmcut approach is (currently) an heuristic approach, if no landmark is found we need to complement it with an exact approach when
+        // needed
         usercuts_lm += cand_cuts::add_comp_lm_cut(context, inst, unreachable_actions, unused_actions, reachable_state);
-    if (exec.cand_cuts.find('s') != std::string::npos)
+    }
+    if (exec.cand_cuts.find('s') != std::string::npos) {
         usercuts_sec += cand_cuts::add_sec_cut(context, inst, unreachable_actions, used_first_achievers);
+    }
 
     cand_time += GET_TIME() - start_time;
 }