+++ code/cut_separators/relax_callback.hpp
@@ -10,7 +10,7 @@
 #include <cplex.h>
 
 #include "../domain/hplus_algs.hpp"
-#include "../utils/algorithms.hpp"
+#include "../utils/cycle_det.hpp"
 
 namespace callbacks {
 
@@ -18,7 +18,7 @@ namespace callbacks {
  * Thread-specific data for CPLEX callbacks
  */
 struct thread_data {
-    unsigned int usercuts_lm, usercuts_sec, relax_calls, cand_calls;
+    unsigned int usercuts_lm, usercuts_sec, relax_calls, cand_calls, acts_in_lm, n_lm;
     double cand_time, relax_time;
 };
 
@@ -39,13 +39,15 @@ std::unordered_map<std::pair<unsigned int, unsigned int>, double, pair_hash> rel
  * Compute the violated landmark (if there's one) out of the relaxed solution
  */
 [[nodiscard]]
-std::pair<bool, std::vector<unsigned int>> get_violated_landmark(const hplus::instance& inst, const std::vector<double>& relax_point);
+std::pair<bool, std::vector<unsigned int>> get_violated_landmark(const hplus::execution& exec, const hplus::instance& inst,
+                                                                 std::vector<double> relax_point, unsigned int& act_in_lm, unsigned int& n_lm);
 
 /**
  * Compute the violated landmark (if there's one) out of the relaxed solution and reject the relaxed solution
  */
 [[nodiscard]]
-unsigned int add_lm_cut(CPXCALLBACKCONTEXTptr context, const hplus::instance& inst, const std::vector<double>& relax_point);
+unsigned int add_lm_cut(CPXCALLBACKCONTEXTptr context, const hplus::execution& exec, const hplus::instance& inst,
+                        const std::vector<double>& relax_point, unsigned int& n_act_in_lm, unsigned int& n_lm);
 
 /**
  * Compute the violated S.E.C. (if there's one) out of the relaxed solution